requirements.txt
torch
torchvision
sentence-transformers
faiss-cpu
pandas
numpy
scikit-learn
Pillow
aiohttp
aiofiles
tqdm
tensorboard
gradio
matplotlib
seaborn
jupyter
ipykernel
pyyaml